+++ apps/astarte_pairing_api/lib/astarte_pairing_api_web/router.ex
@@ -38,4 +38,8 @@ defmodule Astarte.Pairing.APIWeb.Router do
          DeviceController,
          :verify_credentials
   end
+
+  scope "/health", Astarte.Pairing.APIWeb do
+    get "/", HealthController, :show
+  end
 end