+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.11.0-beta.3] - Unreleased
+- [pairing_api] Add health endpoint.
+- [realm_management_api] Add health endpoint.
+- [housekeeping] Add Prometheus instrumenters and exporters.
+- [trigger_engine] Add health endpoint.
+- [housekeeping] Add health endpoint.
+- [realm_management] Add health endpoint.
+- [pairing] Add health endpoint.
+- [data_updater_plant] Add health endpoint.
+
 ## [0.11.0-beta.2] - 2020-01-24
 ### Added
 - [pairing] Add Prometheus instrumenters and exporters.
@@ -16,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [housekeeping_api] Add Prometheus instrumenters and exporters
 - Add standard interfaces for generic sensors.
 - [trigger_engine] Add Prometheus instrumenters and exporters
+- [pairing] Expose registration count and get_credentials count metrics.
 
 ### Changed
 - [realm_management] Handle hyphens in `interface_name`. ([#96](https://github.com/astarte-platform/astarte/issues/96))